* coders/{pcd.c,pcx.c,pdb.c,pict.c,stegano.c,wpg.c,xcf.c}: Warnings reduction. * magick/{channel.c,image.c,unix_port.c,render.c}: Warnings reduction.

+++ coders/pcd.c
@@ -235,7 +235,7 @@ static unsigned int DecodeImage(Image *image,unsigned char *luma,
   unsigned char
     *buffer;
 
-  unsigned long
+  unsigned int
     bits,
     plane,
     pcd_length[3],
